#703 further refactoring
